requirements.txt
fastapi
uvicorn
pillow
mediapipe
numpy
requests
openai
firebase-admin